requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
itsdangerous==2.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
